# RealworldWeb/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "RealWorld Conduit API",
    "description" : "Conduit API documentation",
    "contact" : {
      "name" : "RealWorld",
      "url" : "https://realworld.how"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://api.realworld.io/api"
  } ],
...
  "paths" : {
...
    "/articles/feed" : {
      "get" : {
        "tags" : [ "Articles" ],
        "summary" : "Get recent articles from users you follow",
        "description" : "Get most recent articles from users you follow. Use query parameters to limit. Auth is required",
        "operationId" : "GetArticlesFeed",
        "parameters" : [ {
          "name" : "offset",
          "in" : "query",
          "description" : "The number of items to skip before starting to collect the result set.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "The numbers of items to return.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "integer",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Multiple articles",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_2"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : { }
          },
          "422" : {
            "description" : "Unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            }
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ]
      }
    },
    "/articles" : {
      "get" : {
        "tags" : [ "Articles" ],
        "summary" : "Get recent articles globally",
        "description" : "Get most recent articles globally. Use query parameters to filter results. Auth is optional",
        "operationId" : "GetArticles",
        "parameters" : [ {
          "name" : "tag",
          "in" : "query",
          "description" : "Filter by tag",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "author",
          "in" : "query",
          "description" : "Filter by author (username)",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "favorited",
          "in" : "query",
          "description" : "Filter by favorites of a user (username)",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "description" : "The number of items to skip before starting to collect the result set.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "The numbers of items to return.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "minimum" : 1,
            "type" : "integer",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Multiple articles",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_2"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : { }
          },
          "422" : {
            "description" : "Unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Articles" ],
        "summary" : "Create an article",
        "description" : "Create an article. Auth is required",
        "operationId" : "CreateArticle",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/NewArticleRequest"
        },
        "responses" : {
          "201" : {
            "description" : "Single article",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_201"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : { }
          },
          "422" : {
            "description" : "Unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            }
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "x-codegen-request-body-name" : "article"
      }
    },
    "/articles/{slug}" : {
      "get" : {
        "tags" : [ "Articles" ],
        "summary" : "Get an article",
        "description" : "Get an article. Auth not required",
        "operationId" : "GetArticle",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "description" : "Slug of the article to get",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Single article",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_201"
                }
              }
            }
          },
          "422" : {
            "description" : "Unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Articles" ],
        "summary" : "Update an article",
        "description" : "Update an article. Auth is required",
        "operationId" : "UpdateArticle",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "description" : "Slug of the article to update",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "$ref" : "#/components/requestBodies/UpdateArticleRequest"
        },
        "responses" : {
          "200" : {
            "description" : "Single article",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_201"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : { }
          },
          "422" : {
            "description" : "Unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            }
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "x-codegen-request-body-name" : "article"
      },
      "delete" : {
        "tags" : [ "Articles" ],
        "summary" : "Delete an article",
        "description" : "Delete an article. Auth is required",
        "operationId" : "DeleteArticle",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "description" : "Slug of the article to delete",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "No content",
            "content" : { }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : { }
          },
          "422" : {
            "description" : "Unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            }
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Profile" : {
        "required" : [ "bio", "following", "image", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "bio" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string"
          },
          "following" : {
            "type" : "boolean"
          }
        },
        "example" : {
          "image" : "image",
          "following" : true,
          "bio" : "bio",
          "username" : "username"
        }
      },
      "Article" : {
        "required" : [ "author", "body", "createdAt", "description", "favorited", "favoritesCount", "slug", "tagList", "title", "updatedAt" ],
        "type" : "object",
        "properties" : {
          "slug" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "body" : {
            "type" : "string"
          },
          "tagList" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "favorited" : {
            "type" : "boolean"
          },
          "favoritesCount" : {
            "type" : "integer"
          },
          "author" : {
            "$ref" : "#/components/schemas/Profile"
          }
        },
        "example" : {
          "tagList" : [ "tagList", "tagList" ],
          "createdAt" : "2000-01-23T04:56:07.000+00:00",
          "author" : {
            "image" : "image",
            "following" : true,
            "bio" : "bio",
            "username" : "username"
          },
          "description" : "description",
          "title" : "title",
          "body" : "body",
          "favoritesCount" : 0,
          "slug" : "slug",
          "updatedAt" : "2000-01-23T04:56:07.000+00:00",
          "favorited" : true
        }
      },
      "NewArticle" : {
        "required" : [ "body", "description", "title" ],
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "body" : {
            "type" : "string"
          },
          "tagList" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "UpdateArticle" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "body" : {
            "type" : "string"
          }
        }
      },
...
      "GenericErrorModel" : {
        "required" : [ "errors" ],
        "type" : "object",
        "properties" : {
          "errors" : {
            "$ref" : "#/components/schemas/GenericErrorModel_errors"
          }
        }
      },
...
      "inline_response_200_2" : {
        "required" : [ "articles", "articlesCount" ],
        "type" : "object",
        "properties" : {
          "articles" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Article"
            }
          },
          "articlesCount" : {
            "type" : "integer"
          }
        },
        "example" : {
          "articlesCount" : 6,
          "articles" : [ {
            "tagList" : [ "tagList", "tagList" ],
            "createdAt" : "2000-01-23T04:56:07.000+00:00",
            "author" : {
              "image" : "image",
              "following" : true,
              "bio" : "bio",
              "username" : "username"
            },
            "description" : "description",
            "title" : "title",
            "body" : "body",
            "favoritesCount" : 0,
            "slug" : "slug",
            "updatedAt" : "2000-01-23T04:56:07.000+00:00",
            "favorited" : true
          }, {
            "tagList" : [ "tagList", "tagList" ],
            "createdAt" : "2000-01-23T04:56:07.000+00:00",
            "author" : {
              "image" : "image",
              "following" : true,
              "bio" : "bio",
              "username" : "username"
            },
            "description" : "description",
            "title" : "title",
            "body" : "body",
            "favoritesCount" : 0,
            "slug" : "slug",
            "updatedAt" : "2000-01-23T04:56:07.000+00:00",
            "favorited" : true
          } ]
        }
      },
      "inline_response_201" : {
        "required" : [ "article" ],
        "type" : "object",
        "properties" : {
          "article" : {
            "$ref" : "#/components/schemas/Article"
          }
        },
        "example" : {
          "article" : {
            "tagList" : [ "tagList", "tagList" ],
            "createdAt" : "2000-01-23T04:56:07.000+00:00",
            "author" : {
              "image" : "image",
              "following" : true,
              "bio" : "bio",
              "username" : "username"
            },
            "description" : "description",
            "title" : "title",
            "body" : "body",
            "favoritesCount" : 0,
            "slug" : "slug",
            "updatedAt" : "2000-01-23T04:56:07.000+00:00",
            "favorited" : true
          }
        }
      },
...
      "GenericErrorModel_errors" : {
        "required" : [ "body" ],
        "type" : "object",
        "properties" : {
          "body" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "requestBodies" : {
...
      "NewArticleRequest" : {
        "description" : "Article to create",
        "content" : {
          "application/json" : {
            "schema" : {
              "required" : [ "article" ],
              "type" : "object",
              "properties" : {
                "article" : {
                  "$ref" : "#/components/schemas/NewArticle"
                }
              }
            }
          }
        },
        "required" : true
      },
      "UpdateArticleRequest" : {
        "description" : "Article to update",
        "content" : {
          "application/json" : {
            "schema" : {
              "required" : [ "article" ],
              "type" : "object",
              "properties" : {
                "article" : {
                  "$ref" : "#/components/schemas/UpdateArticle"
                }
              }
            }
          }
        },
        "required" : true
      },
...
    },
    "securitySchemes" : {
      "Token" : {
        "type" : "apiKey",
        "description" : "For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.\n\nA JWT token is generated by the API by either registering via /users or logging in via /users/login.\n\nThe following format must be in the 'Authorization' header :\n\n    Token xxxxxx.yyyyyyy.zzzzzz\n    \n",
        "name" : "Authorization",
        "in" : "header"
      }
    }
  }
}